pyproject.toml
[project]
name = "infraprobe"
version = "0.1.0"
description = "Infrastructure probing API"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.34",
    "cryptography>=46.0.5",
    "dnspython>=2.7",
    "sslyze>=6.3.0",
    "checkdmarc==4.8.5",
    "wappalyzer>=1.0.20",
    "drheaderplus>=3.0.0",
    "python-nmap>=0.7.1",
    "asyncwhois>=1.1.12",
    "prometheus-client>=0.21",
]

[project.optional-dependencies]
firestore = ["google-cloud-firestore>=2.19"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/infraprobe"]

[dependency-groups]
dev = [
    "httpx>=0.28",
    "locust>=2.32",
    "pip-audit>=2.7",
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "pytest-httpserver>=1.1.5",
    "pyyaml>=6.0.3",
    "ruff>=0.9",
    "ty",
]

[tool.uv]
override-dependencies = ["cryptography>=46.0.5"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit real external targets (require network)",
    "slow: tests that take a long time to run",
]
